Split map.go into multiple go files
This change makes the code easier to read and maintain.

This commit splits map.go into:
- map_data_slab.go
- map_data_slab_decode.go
- map_data_slab_encode.go
- map_element.go
- map_element_decode.go
- map_element_encode.go
- map_elements.go
- map_elements_decode.go
- map_elements_encode.go
- map_elements_hashkey.go
- map_elements_nokey.go
- map_extradata.go
- map_iterator.go
- map_metadata_slab.go
- map_metadata_slab_decode.go
- map_metadata_slab_encode.go
- map_size_consts.go
- map_slab.go

While at it, this commit also groups related functions together.
